[refactor] move docker related utils to its own file (#1456)
* move docker utils to its own
