+++ tests/yatai/local_yatai_service.py
@@ -9,7 +9,7 @@ import docker
 
 from bentoml.configuration import LAST_PYPI_RELEASE_VERSION
 from bentoml.utils.tempdir import TempDirectory
-from bentoml.yatai.deployment.utils import ensure_docker_available_or_raise
+from bentoml.yatai.deployment.docker_utils import ensure_docker_available_or_raise
 
 logger = logging.getLogger('bentoml.test')
